+++ CHANGELOG.md
@@ -5,10 +5,17 @@
 ### Features
 
 - Add 24 Hour Volume, Latest Price and Market Cap to the dashboard #77
+- Support query params for wallet to auto login #98
+- Add UI to claim rewards in staking #164
+
+### Improvements
+
+- Table lists pull more than 10 results #176
 
 ### Bug Fixes
 
 - Fix for mobile overflow #68
+- Fix for validator images #180
 
 ## 2.1.1 - 2021-07-30